try to create a unified logging interface| MLflow
