+++ pycaret/internal/tabular.py
@@ -6,6 +6,7 @@
 
 from enum import Enum, auto
 import math
+from pycaret import loggers
 from pycaret.internal.meta_estimators import (
     PowerTransformedTargetRegressor,
     CustomProbabilityThresholdClassifier,
@@ -41,6 +42,7 @@ from pycaret.internal.drift_report import (
     create_classification_drift_report,
     create_regression_drift_report,
 )
+from pycaret.loggers.mlflow_logger import mlflowLogger
 import pycaret.containers.metrics.classification
 import pycaret.containers.metrics.regression
 import pycaret.containers.metrics.clustering
@@ -793,7 +795,7 @@ def setup(
     # declaring global variables to be accessed by other functions
     logger.info("Declaring global variables")
     global _ml_usecase, USI, html_param, X, y, X_train, X_test, y_train, y_test, seed, prep_pipe, experiment__, fold_shuffle_param, n_jobs_param, _gpu_n_jobs_param, create_model_container, master_model_container, display_container, exp_name_log, logging_param, log_plots_param, fix_imbalance_param, fix_imbalance_method_param, transform_target_param, transform_target_method_param, data_before_preprocess, target_param, gpu_param, _all_models, _all_models_internal, _all_metrics, _internal_pipeline, stratify_param, fold_generator, fold_param, fold_groups_param, fold_groups_param_full, imputation_regressor, imputation_classifier, iterative_imputation_iters_param
-
+    global dashboard_logger
     USI = secrets.token_hex(nbytes=2)
     logger.info(f"USI: {USI}")
 
@@ -845,6 +847,7 @@ def setup(
         "fold_param",
         "fold_groups_param",
         "fold_groups_param_full",
+        "dashboard_logger"
     }
     if not _is_unsupervised(_ml_usecase):
         pycaret_globals = common_globals.union(supervised_globals)
@@ -1294,6 +1297,9 @@ def setup(
 
     # create logging parameter
     logging_param = log_experiment
+    if logging_param:
+        loggers_list = [mlflowLogger()]
+        dashboard_logger = loggers.DashboardLogger(loggers_list)
 
     # create exp_name_log param incase logging is False
     exp_name_log = "no_logging"
@@ -1745,81 +1751,6 @@ def setup(
     runtime = np.array(runtime_end - runtime_start).round(2)
 
     if logging_param:
-
-        logger.info("Logging experiment in MLFlow")
-
-        import mlflow
-
-        try:
-            mlflow.create_experiment(exp_name_log)
-        except:
-            logger.warning("Couldn't create mlflow experiment. Exception:")
-            logger.warning(traceback.format_exc())
-
-        # mlflow logging
-        mlflow.set_experiment(exp_name_log)
-
-        run_name_ = f"Session Initialized {USI}"
-
-        mlflow.end_run()
-        mlflow.start_run(run_name=run_name_)
-
-        # Get active run to log as tag
-        RunID = mlflow.active_run().info.run_id
-
-        k = functions.copy()
-        k.set_index("Description", drop=True, inplace=True)
-        kdict = k.to_dict()
-        params = kdict.get("Value")
-        mlflow.log_params(params)
-
-        # set tag of compare_models
-        mlflow.set_tag("Source", "setup")
-
-        # set custom tags if applicable
-        if isinstance(experiment_custom_tags, dict):
-            mlflow.set_tags(experiment_custom_tags)
-
-        import secrets
-
-        URI = secrets.token_hex(nbytes=4)
-        mlflow.set_tag("URI", URI)
-        mlflow.set_tag("USI", USI)
-        mlflow.set_tag("Run Time", runtime)
-        mlflow.set_tag("Run ID", RunID)
-
-        # Log the transformation pipeline
-        logger.info("SubProcess save_model() called ==================================")
-        save_model(prep_pipe, "Transformation Pipeline", verbose=False)
-        logger.info("SubProcess save_model() end ==================================")
-        mlflow.log_artifact("Transformation Pipeline.pkl")
-        os.remove("Transformation Pipeline.pkl")
-
-        # Log pandas profile
-        if log_profile:
-            import pandas_profiling
-
-            pf = pandas_profiling.ProfileReport(
-                data_before_preprocess, **profile_kwargs
-            )
-            pf.to_file("Data Profile.html")
-            mlflow.log_artifact("Data Profile.html")
-            os.remove("Data Profile.html")
-            display.display(functions_, clear=True)
-
-        # Log training and testing set
-        if log_data:
-            if not _is_unsupervised(_ml_usecase):
-                X_train.join(y_train).to_csv("Train.csv")
-                X_test.join(y_test).to_csv("Test.csv")
-                mlflow.log_artifact("Train.csv")
-                mlflow.log_artifact("Test.csv")
-                os.remove("Train.csv")
-                os.remove("Test.csv")
-            else:
-                X.to_csv("Dataset.csv")
-                mlflow.log_artifact("Dataset.csv")
-                os.remove("Dataset.csv")
 
     logger.info(f"create_model_container: {len(create_model_container)}")
     logger.info(f"master_model_container: {len(master_model_container)}")
@@ -2389,9 +2320,10 @@ def compare_models(
             if logging_param and cross_validation and model is not None:
 
                 try:
-                    _mlflow_log_model(
+                    mlflowLogger.log_model(
                         model=model,
                         model_results=results,
+                        ml_usecase=_ml_usecase,
                         score_dict=avgs_dict_log,
                         source="compare_models",
                         runtime=row["runtime"],
@@ -2756,9 +2688,10 @@ def create_model_unsupervised(
         metrics_log = {k: v for k, v in metrics.items()}
 
         try:
-            _mlflow_log_model(
+            mlflowLogger.log_model(
                 model=model,
                 model_results=None,
+                ml_usecase=_ml_usecase,
                 score_dict=metrics_log,
                 source="create_model",
                 experiment_custom_tags=experiment_custom_tags,
@@ -3274,9 +3207,10 @@ def create_model_supervised(
         avgs_dict_log = {k: v[0] for k, v in avgs_dict_log.items()}
 
         try:
-            _mlflow_log_model(
+            mlflowLogger.log_model(
                 model=model,
                 model_results=model_results,
+                ml_usecase=_ml_usecase,
                 score_dict=avgs_dict_log,
                 source="create_model",
                 runtime=runtime,
@@ -3615,9 +3549,10 @@ def tune_model_unsupervised(
         metrics_log = {k: v[0] for k, v in best_model_results.items()}
 
         try:
-            _mlflow_log_model(
+            mlflowLogger.log_model(
                 model=model,
                 model_results=None,
+                ml_usecase=_ml_usecase,
                 score_dict=metrics_log,
                 source="tune_model",
                 runtime=runtime,
@@ -4663,9 +4598,10 @@ def tune_model_supervised(
         avgs_dict_log = {k: v for k, v in model_results.loc["Mean"].items()}
 
         try:
-            _mlflow_log_model(
+            mlflowLogger.log_model(
                 model=best_model,
                 model_results=model_results,
+                ml_usecase=_ml_usecase,
                 score_dict=avgs_dict_log,
                 source="tune_model",
                 runtime=runtime,
@@ -5009,9 +4945,10 @@ def ensemble_model(
         avgs_dict_log = {k: v for k, v in model_results.loc["Mean"].items()}
 
         try:
-            _mlflow_log_model(
+            mlflowLogger.log_model(
                 model=best_model,
                 model_results=model_results,
+                ml_usecase=_ml_usecase,
                 score_dict=avgs_dict_log,
                 source="ensemble_model",
                 runtime=runtime,
@@ -5357,9 +5294,10 @@ def blend_models(
         avgs_dict_log = {k: v for k, v in model_results.loc["Mean"].items()}
 
         try:
-            _mlflow_log_model(
+            mlflowLogger.log_model(
                 model=model,
                 model_results=model_results,
+                ml_usecase=_ml_usecase,
                 score_dict=avgs_dict_log,
                 source="blend_models",
                 runtime=runtime,
@@ -5716,9 +5654,10 @@ def stack_models(
         avgs_dict_log = {k: v for k, v in model_results.loc["Mean"].items()}
 
         try:
-            _mlflow_log_model(
+            mlflowLogger.log_model(
                 model=model,
                 model_results=model_results,
+                ml_usecase=_ml_usecase,
                 score_dict=avgs_dict_log,
                 source="stack_models",
                 runtime=runtime,
@@ -8309,9 +8248,10 @@ def calibrate_model(
         avgs_dict_log = {k: v for k, v in model_results.loc["Mean"].items()}
 
         try:
-            _mlflow_log_model(
+            mlflowLogger.log_model(
                 model=model,
                 model_results=model_results,
+                ml_usecase=_ml_usecase,
                 score_dict=avgs_dict_log,
                 source="calibrate_models",
                 runtime=runtime,
@@ -9042,9 +8982,10 @@ def finalize_model(
         avgs_dict_log = {k: v for k, v in model_results.loc["Mean"].items()}
 
         try:
-            _mlflow_log_model(
+            mlflowLogger.log_model(
                 model=model_final,
                 model_results=model_results,
+                ml_usecase=_ml_usecase,
                 score_dict=avgs_dict_log,
                 source="finalize_model",
                 runtime=runtime,
@@ -10917,196 +10858,6 @@ def get_ml_task(y):
     else:
         ml_usecase = MLUsecase.REGRESSION
     return ml_usecase
-
-
-def _mlflow_log_model(
-    model,
-    model_results,
-    score_dict: dict,
-    source: str,
-    runtime: float,
-    model_fit_time: float,
-    _prep_pipe,
-    log_holdout: bool = True,
-    log_plots: bool = False,
-    experiment_custom_tags: Optional[Dict[str, Any]] = None,
-    tune_cv_results=None,
-    URI=None,
-    display: Optional[Display] = None,
-):
-    logger = get_logger()
-
-    logger.info("Creating MLFlow logs")
-
-    # Creating Logs message monitor
-    if display:
-        display.update_monitor(1, "Creating Logs")
-        display.display_monitor()
-
-    # import mlflow
-    import mlflow
-    import mlflow.sklearn
-
-    mlflow.set_experiment(exp_name_log)
-
-    full_name = _get_model_name(model)
-    logger.info(f"Model: {full_name}")
-
-    with mlflow.start_run(run_name=full_name, nested=True) as run:
-
-        # Get active run to log as tag
-        RunID = mlflow.active_run().info.run_id
-
-        # Log model parameters
-        pipeline_estimator_name = get_pipeline_estimator_label(model)
-        if pipeline_estimator_name:
-            params = model.named_steps[pipeline_estimator_name]
-        else:
-            params = model
-
-        # get regressor from meta estimator
-        params = get_estimator_from_meta_estimator(params)
-
-        try:
-            try:
-                params = params.get_all_params()
-            except:
-                params = params.get_params()
-        except:
-            logger.warning("Couldn't get params for model. Exception:")
-            logger.warning(traceback.format_exc())
-            params = {}
-
-        for i in list(params):
-            v = params.get(i)
-            if len(str(v)) > 250:
-                params.pop(i)
-
-        logger.info(f"logged params: {params}")
-        mlflow.log_params(params)
-
-        # Log metrics
-        mlflow.log_metrics(score_dict)
-
-        # set tag of compare_models
-        mlflow.set_tag("Source", source)
-
-        # set custom tags if applicable
-        if isinstance(experiment_custom_tags, dict):
-            mlflow.set_tags(experiment_custom_tags)
-
-        if not URI:
-            import secrets
-
-            URI = secrets.token_hex(nbytes=4)
-        mlflow.set_tag("URI", URI)
-        mlflow.set_tag("USI", USI)
-        mlflow.set_tag("Run Time", runtime)
-        mlflow.set_tag("Run ID", RunID)
-
-        # Log training time in seconds
-        mlflow.log_metric("TT", model_fit_time)
-
-        # Log the CV results as model_results.html artifact
-        if not _is_unsupervised(_ml_usecase):
-            try:
-                model_results.data.to_html("Results.html", col_space=65, justify="left")
-            except:
-                model_results.to_html("Results.html", col_space=65, justify="left")
-            mlflow.log_artifact("Results.html")
-            os.remove("Results.html")
-
-            if log_holdout:
-                # Generate hold-out predictions and save as html
-                try:
-                    holdout = predict_model(model, verbose=False)
-                    holdout_score = pull(pop=True)
-                    del holdout
-                    holdout_score.to_html("Holdout.html", col_space=65, justify="left")
-                    mlflow.log_artifact("Holdout.html")
-                    os.remove("Holdout.html")
-                except:
-                    logger.warning(
-                        "Couldn't create holdout prediction for model, exception below:"
-                    )
-                    logger.warning(traceback.format_exc())
-
-        # Log AUC and Confusion Matrix plot
-
-        if log_plots:
-
-            logger.info(
-                "SubProcess plot_model() called =================================="
-            )
-
-            def _log_plot(plot):
-                try:
-                    plot_name = plot_model(
-                        model, plot=plot, verbose=False, save=True, system=False
-                    )
-                    mlflow.log_artifact(plot_name)
-                    os.remove(plot_name)
-                except Exception as e:
-                    logger.warning(e)
-
-            for plot in log_plots:
-                _log_plot(plot)
-
-            logger.info(
-                "SubProcess plot_model() end =================================="
-            )
-
-        # Log hyperparameter tuning grid
-        if tune_cv_results:
-            d1 = tune_cv_results.get("params")
-            dd = pd.DataFrame.from_dict(d1)
-            dd["Score"] = tune_cv_results.get("mean_test_score")
-            dd.to_html("Iterations.html", col_space=75, justify="left")
-            mlflow.log_artifact("Iterations.html")
-            os.remove("Iterations.html")
-
-        # get default conda env
-        from mlflow.sklearn import get_default_conda_env
-
-        default_conda_env = get_default_conda_env()
-        default_conda_env["name"] = f"{exp_name_log}-env"
-        default_conda_env.get("dependencies").pop(-3)
-        dependencies = default_conda_env.get("dependencies")[-1]
-        from pycaret.utils import __version__
-
-        dep = f"pycaret=={__version__}"
-        dependencies["pip"] = [dep]
-
-        # define model signature
-        from mlflow.models.signature import infer_signature
-
-        try:
-            signature = infer_signature(
-                data_before_preprocess.drop([target_param], axis=1)
-            )
-        except:
-            logger.warning("Couldn't infer MLFlow signature.")
-            signature = None
-        if not _is_unsupervised(_ml_usecase):
-            input_example = (
-                data_before_preprocess.drop([target_param], axis=1).iloc[0].to_dict()
-            )
-        else:
-            input_example = data_before_preprocess.iloc[0].to_dict()
-
-        # log model as sklearn flavor
-        prep_pipe_temp = deepcopy(_prep_pipe)
-        prep_pipe_temp.steps.append(["trained_model", model])
-        mlflow.sklearn.log_model(
-            prep_pipe_temp,
-            "model",
-            conda_env=default_conda_env,
-            # signature=signature,
-            # input_example=input_example,
-        )
-        del prep_pipe_temp
-    gc.collect()
-
 
 def _get_columns_to_stratify_by(
     X: pd.DataFrame, y: pd.DataFrame, stratify: Union[bool, List[str]], target: str
