+++ pycaret/loggers/wandb_logger.py
@@ -1,0 +1,1 @@
+from pycaret.loggers import BaseLogger
